(fix) make time column in transaction an integer

+++ merkato/utils/database_utils.py
@@ -364,7 +364,7 @@ def create_transactions_table():
     finally:
         c = conn.cursor()
         c.execute('''CREATE TABLE IF NOT EXISTS transactions
-                    (uuid text, base text, quote text, spread float, tx_id text, order_id text, price float, amount float, time float)''')
+                    (uuid text, base text, quote text, spread float, tx_id text, order_id text, price float, amount float, time integer)''')
         c.execute('''CREATE UNIQUE INDEX id_exchange_pair ON transactions (uuid)''')
         conn.commit()
         conn.close()
